+++ MvvmPlugin/MvvmPlugin/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.5.0
+- Added Go to view context action
+- Fix issue with navigating to newly created view model when calling Create ViewModel context action
+
 ## 0.4.1
 - Fixed critical bug with macro